Move some methods to the defaults flavor only
