+++ api-tester/src/defaults/kotlin/com/revenuecat/apitester/kotlin/PurchasesAPI.kt
@@ -11,11 +11,13 @@ import com.revenuecat.purchases.PurchaseParams
 import com.revenuecat.purchases.Purchases
 import com.revenuecat.purchases.PurchasesConfiguration
 import com.revenuecat.purchases.PurchasesError
+import com.revenuecat.purchases.Store
 import com.revenuecat.purchases.awaitCustomerInfo
 import com.revenuecat.purchases.getCustomerInfoWith
 import com.revenuecat.purchases.interfaces.LogInCallback
 import com.revenuecat.purchases.interfaces.PurchaseCallback
 import com.revenuecat.purchases.interfaces.ReceiveCustomerInfoCallback
+import com.revenuecat.purchases.interfaces.SyncPurchasesCallback
 import com.revenuecat.purchases.logInWith
 import com.revenuecat.purchases.logOutWith
 import com.revenuecat.purchases.models.BillingFeature
@@ -23,10 +25,12 @@ import com.revenuecat.purchases.models.GoogleProrationMode
 import com.revenuecat.purchases.models.StoreProduct
 import com.revenuecat.purchases.models.StoreTransaction
 import com.revenuecat.purchases.models.SubscriptionOption
+import com.revenuecat.purchases.restorePurchasesWith
+import com.revenuecat.purchases.syncPurchasesWith
 import java.util.concurrent.ExecutorService
 
 @OptIn(ExperimentalPreviewRevenueCatPurchasesAPI::class)
-@Suppress("unused", "UNUSED_VARIABLE", "EmptyFunctionBlock")
+@Suppress("unused", "UNUSED_VARIABLE", "EmptyFunctionBlock", "DEPRECATION")
 private class PurchasesAPI {
     @SuppressWarnings("LongParameterList")
     fun check(
@@ -40,18 +44,33 @@ private class PurchasesAPI {
             override fun onReceived(customerInfo: CustomerInfo, created: Boolean) {}
             override fun onError(error: PurchasesError) {}
         }
+        val syncPurchasesCallback = object : SyncPurchasesCallback {
+            override fun onSuccess(customerInfo: CustomerInfo) {}
+            override fun onError(error: PurchasesError) {}
+        }
+
+        purchases.syncPurchases()
+        purchases.syncPurchases(syncPurchasesCallback)
 
         purchases.logIn("", logInCallback)
         purchases.logOut()
         purchases.logOut(receiveCustomerInfoCallback)
 
+        purchases.restorePurchases(receiveCustomerInfoCallback)
+        purchases.invalidateCustomerInfoCache()
+
         purchases.getCustomerInfo(receiveCustomerInfoCallback)
         purchases.getCustomerInfo(CacheFetchPolicy.CACHED_OR_FETCHED, receiveCustomerInfoCallback)
+
+        val finishTransactions: Boolean = purchases.finishTransactions
+        purchases.finishTransactions = true
 
         val anonymous: Boolean = purchases.isAnonymous
 
         purchases.onAppBackgrounded()
         purchases.onAppForegrounded()
+
+        val store: Store = purchases.store
     }
 
     @SuppressWarnings("LongParameterList", "EmptyFunctionBlock")
@@ -107,6 +126,17 @@ private class PurchasesAPI {
         purchases.getCustomerInfoWith(
             fetchPolicy = CacheFetchPolicy.CACHED_OR_FETCHED,
             onError = { _: PurchasesError -> },
+            onSuccess = { _: CustomerInfo -> },
+        )
+        purchases.restorePurchasesWith(
+            onError = { _: PurchasesError -> },
+            onSuccess = { _: CustomerInfo -> },
+        )
+        purchases.syncPurchasesWith(
+            onError = { _: PurchasesError -> },
+            onSuccess = { _: CustomerInfo -> },
+        )
+        purchases.syncPurchasesWith(
             onSuccess = { _: CustomerInfo -> },
         )
     }
